(DBFLUTE-1030){Java/C#}: LoadDataReverseを繰り返し利用しやすいように改善

+++ dbflute/src/main/java/org/seasar/dbflute/logic/doc/lreverse/DfLReverseProcess.java
@@ -254,13 +254,15 @@ public class DfLReverseProcess {
                 return targetPrefix && name.endsWith(".xls");
             }
         });
-        final String delimiterDataDir = _outputHandler.getDelimiterDataDir();
-        if (delimiterDataDir != null) {
-            doDeletePreviousDataFile(new File(delimiterDataDir), new FileFilter() {
-                public boolean accept(File file) {
-                    return file.getName().endsWith(".tsv");
-                }
-            });
+        if (!isReplaceSchemaDirectUse()) { // not to remove non-related TSV files
+            final String delimiterDataDir = _outputHandler.getDelimiterDataDir();
+            if (delimiterDataDir != null) {
+                doDeletePreviousDataFile(new File(delimiterDataDir), new FileFilter() {
+                    public boolean accept(File file) {
+                        return file.getName().endsWith(".tsv");
+                    }
+                });
+            }
         }
     }
